+++ aidial_sdk/chat_completion/request.py
@@ -8,8 +8,7 @@ from typing_extensions import Annotated, assert_never
 from aidial_sdk.chat_completion.enums import Status
 from aidial_sdk.deployment.from_request_mixin import FromRequestDeploymentMixin
 from aidial_sdk.exceptions import InvalidRequestError
-from aidial_sdk.pydantic._compat import PYDANTIC_V2
-from aidial_sdk.pydantic._models import model_validator
+from aidial_sdk.pydantic._compat import PYDANTIC_V2, model_validator
 from aidial_sdk.utils.pydantic import ExtraForbidModel
