+++ aidial_sdk/deployment/tokenize.py
@@ -2,7 +2,7 @@ from typing import List, Literal, Union
 
 from aidial_sdk.chat_completion.request import ChatCompletionRequest
 from aidial_sdk.deployment.from_request_mixin import FromRequestDeploymentMixin
-from aidial_sdk.pydantic._models import BaseModel
+from aidial_sdk.pydantic._compat import BaseModel
 
 
 class TokenizeInputRequest(BaseModel):
